write to the Thrall Kinesis stream as gzip
Grid reads messages from Kinesis by removing the first byte (the compression marker) and gunzipping it.

RCS Poller Lambda hasn't been updated to write to Kinesis in this format. That is, RCS Poller Lambda is writing plain JSON to Kinesis and then Thrall is failing to gunzip it and dropping the message.

This updates RCS Poller Lambda to write to the Thrall Kinesis stream in a valid format: leading byte is 0 and the remaining data is gzip compressed.